Crypto.PBKDF

Password Based Key Derivation Functions

This module provides stock implementations of the PBKDF functions from RFC-2898 (http://www.ietf.org/rfc/rfc2898.txt), based on the SHA-1, SHA-256 and SHA-256 hash functions. Each function takes the password and salt as a string and returns a hex string. To work with ByteStrings see Crypto.PBKDF.ByteString.

SHA-based PBKDF1 functions

sha1PBKDF1

Arguments

:: String
the password (will be encoded with UTF-8)

-> String
the salt (will be encoded with UTF-8)

-> Int
the iteration count

-> String
the result key as a hex string

sha256PBKDF1

Arguments

:: String
the password (will be encoded with UTF-8)

-> String
the salt (will be encoded with UTF-8)

-> Int
the iteration count

-> String
the result key as a hex string

sha512PBKDF1

Arguments

:: String
the password (will be encoded with UTF-8)

-> String
the salt (will be encoded with UTF-8)

-> Int
the iteration count

-> String
the result key as a hex string

SHA-based PBKDF2 functions

sha1PBKDF2

Arguments

:: String
the password (will be encoded with UTF-8)

-> String
the salt (will be encoded with UTF-8)

-> Int
the iteration count

-> Int
the length of the key to be generated (in octets)

-> String
the result key as a hex string

sha256PBKDF2

Arguments

:: String
the password (will be encoded with UTF-8)

-> String
the salt (will be encoded with UTF-8)

-> Int
the iteration count

-> Int
the length of the key to be generated (in octets)

-> String
the result key as a hex string

sha512PBKDF2

Arguments

:: String
the password (will be encoded with UTF-8)

-> String
the salt (will be encoded with UTF-8)

-> Int
the iteration count

-> Int
the length of the key to be generated (in octets)

-> String
the result key as a hex string